+++ packages/eslint-plugin-template/src/rules/click-events-have-key-events.ts
@@ -6,7 +6,7 @@ import { isHiddenFromScreenReader } from '../utils/is-hidden-from-screen-reader'
 import { isInteractiveElement } from '../utils/is-interactive-element';
 import { isPresentationRole } from '../utils/is-presentation-role';
 
-type Options = [];
+export type Options = [];
 export type MessageIds = 'clickEventsHaveKeyEvents';
 export const RULE_NAME = 'click-events-have-key-events';