+++ packages/eslint-plugin-template/src/rules/table-scope.ts
@@ -7,7 +7,7 @@ import { createESLintRule } from '../utils/create-eslint-rule';
 import { getDomElements } from '../utils/get-dom-elements';
 import { toPattern } from '../utils/to-pattern';
 
-type Options = [];
+export type Options = [];
 export type MessageIds = 'tableScope';
 export const RULE_NAME = 'table-scope';