+++ packages/eslint-plugin-template/tests/rules/label-has-associated-control/cases.ts
@@ -1,9 +1,16 @@
 import { convertAnnotatedSourceToFailureCase } from '@angular-eslint/test-utils';
-import type { MessageIds } from '../../../src/rules/label-has-associated-control';
+import type {
+  InvalidTestCase,
+  ValidTestCase,
+} from '@typescript-eslint/rule-tester';
+import type {
+  MessageIds,
+  Options,
+} from '../../../src/rules/label-has-associated-control';
 
 const messageId: MessageIds = 'labelHasAssociatedControl';
 
-export const valid = [
+export const valid: readonly (string | ValidTestCase<Options>)[] = [
   `
     <ng-container *ngFor="let item of items; index as index">
       <label for="item-{{index}}">Label #{{index}</label>
@@ -67,7 +74,7 @@ export const valid = [
   },
 ];
 
-export const invalid = [
+export const invalid: readonly InvalidTestCase<MessageIds, Options>[] = [
   convertAnnotatedSourceToFailureCase({
     messageId,
     description: 'should fail if a label does not have a "for" attribute',