+++ packages/eslint-plugin/src/rules/use-component-selector.ts
@@ -2,7 +2,7 @@ import { ASTUtils, Selectors } from '@angular-eslint/utils';
 import type { TSESTree } from '@typescript-eslint/utils';
 import { createESLintRule } from '../utils/create-eslint-rule';
 
-type Options = [];
+export type Options = [];
 export type MessageIds = 'useComponentSelector';
 export const RULE_NAME = 'use-component-selector';