+++ packages/eslint-plugin/tests/rules/no-inputs-metadata-property/cases.ts
@@ -1,9 +1,16 @@
 import { convertAnnotatedSourceToFailureCase } from '@angular-eslint/test-utils';
-import type { MessageIds } from '../../../src/rules/no-inputs-metadata-property';
+import type {
+  InvalidTestCase,
+  ValidTestCase,
+} from '@typescript-eslint/rule-tester';
+import type {
+  MessageIds,
+  Options,
+} from '../../../src/rules/no-inputs-metadata-property';
 
 const messageId: MessageIds = 'noInputsMetadataProperty';
 
-export const valid = [
+export const valid: readonly (string | ValidTestCase<Options>)[] = [
   `class Test {}`,
   `
     @Component()
@@ -82,7 +89,7 @@ export const valid = [
  `,
 ];
 
-export const invalid = [
+export const invalid: readonly InvalidTestCase<MessageIds, Options>[] = [
   convertAnnotatedSourceToFailureCase({
     description:
       'should fail if `inputs` metadata property is used in `@Component`',